Update route of login form
  - Updating file path of the login form
